pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pykasacloud"
dynamic = ["version"]
authors = [
  { name="iluvdata", email="gbellsworth@gmail.com" },
]
description = "Unofficial library for interacting with Kasa cloud api."
readme = "readme.md"
requires-python = ">=3.9"
classifiers=[
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Home Automation",
]
dependencies = [
  "python-kasa==0.10.2"
]

[project.urls]
Homepage = "https://github.com/iluvdata/PyKasaCloud"
Issues = "https://github.com/iluvdata/PyKasaCloud/issues"

[tool.hatch.version]
path = "pykasacloud/__version__.py"